refactor: use convert-units for conversion
